fix pytests
